factor out comfy-table
