+++ src/common/display/Cargo.toml
@@ -1,4 +1,5 @@
 [dependencies]
+comfy-table = {workspace = true}
 indexmap = {workspace = true}
 pyo3 = {workspace = true, optional = true}
 terminal_size = {version = "0.3.0"}
