+++ src/common/display/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod ascii;
 pub mod mermaid;
+pub mod table_display;
 pub mod tree;
 pub mod utils;
