+++ src/daft-schema/Cargo.toml
@@ -1,6 +1,5 @@
 [dependencies]
 arrow2 = {workspace = true}
-comfy-table = {workspace = true}
 common-arrow-ffi = {path = "../common/arrow-ffi", default-features = false}
 common-display = {path = "../common/display", default-features = false}
 common-error = {path = "../common/error", default-features = false}
