+++ src/daft-schema/src/schema.rs
@@ -1,19 +1,18 @@
 use std::{
-    borrow::Cow,
     collections::{hash_map::DefaultHasher, HashSet},
     fmt::{Display, Formatter, Result},
     hash::{Hash, Hasher},
     sync::Arc,
 };
 
-use common_display::DisplayAs;
+use common_display::{
+    table_display::{make_schema_horizontal_table, make_schema_vertical_table},
+    DisplayAs,
+};
 use indexmap::IndexMap;
 use serde::{Deserialize, Serialize};
 
-use crate::{
-    field::Field,
-    // utils::display_table::{make_comfy_table, make_schema_vertical_table},
-};
+use crate::field::Field;
 
 use common_error::{DaftError, DaftResult};
 
@@ -219,7 +218,7 @@ impl Schema {
         let table = make_schema_horizontal_table(
             self.fields
                 .values()
-                .map(Cow::Borrowed)
+                .map(|field| format!("{}\n---\n{}", field.name, field.dtype))
                 .collect::<Vec<_>>()
                 .as_slice(),
         );
@@ -289,13 +288,17 @@ impl Default for Schema {
 impl Display for Schema {
     // Produces an ASCII table.
     fn fmt(&self, f: &mut Formatter) -> Result {
-        let table = make_schema_vertical_table(
-            self.fields
-                .values()
-                .map(Cow::Borrowed)
-                .collect::<Vec<_>>()
-                .as_slice(),
-        );
+        let names = self
+            .fields
+            .values()
+            .map(|f| f.name.clone())
+            .collect::<Vec<_>>();
+        let dtypes = self
+            .fields
+            .values()
+            .map(|f| format!("{}", f.dtype))
+            .collect::<Vec<_>>();
+        let table = make_schema_vertical_table(&names, &dtypes);
         writeln!(f, "{table}")
     }
 }
@@ -317,112 +320,4 @@ impl TryFrom<&arrow2::datatypes::Schema> for Schema {
         let daft_fields: Vec<Field> = fields.iter().map(|f| f.into()).collect();
         Self::new(daft_fields)
     }
-}
-const BOLD_TABLE_HEADERS_IN_DISPLAY: &str = "DAFT_BOLD_TABLE_HEADERS";
-
-// this should be factored out to a common crate
-fn create_table_cell(value: &str) -> comfy_table::Cell {
-    let mut attributes = vec![];
-    if std::env::var(BOLD_TABLE_HEADERS_IN_DISPLAY)
-        .as_deref()
-        .unwrap_or("1")
-        == "1"
-    {
-        attributes.push(comfy_table::Attribute::Bold);
-    }
-
-    let mut cell = comfy_table::Cell::new(value);
-    if !attributes.is_empty() {
-        cell = cell.add_attributes(attributes);
-    }
-    cell
-}
-
-fn make_schema_vertical_table<F: AsRef<Field>>(fields: &[F]) -> comfy_table::Table {
-    let mut table = comfy_table::Table::new();
-
-    let default_width_if_no_tty = 120usize;
-
-    table
-        .load_preset(comfy_table::presets::UTF8_FULL)
-        .apply_modifier(comfy_table::modifiers::UTF8_ROUND_CORNERS)
-        .set_content_arrangement(comfy_table::ContentArrangement::Dynamic);
-    if table.width().is_none() && !table.is_tty() {
-        table.set_width(default_width_if_no_tty as u16);
-    }
-
-    let header = vec![create_table_cell("Column Name"), create_table_cell("Type")];
-    table.set_header(header);
-
-    for f in fields.iter() {
-        table.add_row(vec![
-            f.as_ref().name.to_string(),
-            format!("{}", f.as_ref().dtype),
-        ]);
-    }
-    table
-}
-
-fn make_schema_horizontal_table<F: AsRef<Field>>(fields: &[F]) -> comfy_table::Table {
-    let mut table = comfy_table::Table::new();
-
-    let default_width_if_no_tty = 120usize;
-
-    table
-        .load_preset(comfy_table::presets::UTF8_FULL)
-        .apply_modifier(comfy_table::modifiers::UTF8_ROUND_CORNERS)
-        .set_content_arrangement(comfy_table::ContentArrangement::Dynamic);
-    if table.width().is_none() && !table.is_tty() {
-        table.set_width(default_width_if_no_tty as u16);
-    }
-    let terminal_width = table
-        .width()
-        .expect("should have already been set with default") as usize;
-
-    let expected_col_width = 18usize;
-
-    let max_cols = (((terminal_width + expected_col_width - 1) / expected_col_width) - 1).max(1);
-    const DOTS: &str = "…";
-    let num_columns = fields.len();
-
-    let head_cols;
-    let tail_cols;
-    if num_columns > max_cols {
-        head_cols = (max_cols + 1) / 2;
-        tail_cols = max_cols / 2;
-    } else {
-        head_cols = num_columns;
-        tail_cols = 0;
-    }
-    let mut header = fields
-        .iter()
-        .take(head_cols)
-        .map(|field| {
-            create_table_cell(&format!(
-                "{}\n---\n{}",
-                field.as_ref().name,
-                field.as_ref().dtype
-            ))
-        })
-        .collect::<Vec<_>>();
-    if tail_cols > 0 {
-        let unseen_cols = num_columns - (head_cols + tail_cols);
-        header.push(
-            create_table_cell(&format!(
-                "{DOTS}\n\n({unseen_cols} hidden)",
-                DOTS = DOTS,
-                unseen_cols = unseen_cols
-            ))
-            .set_alignment(comfy_table::CellAlignment::Center),
-        );
-        header.extend(fields.iter().skip(num_columns - tail_cols).map(|field| {
-            create_table_cell(&format!(
-                "{}\n---\n{}",
-                field.as_ref().name,
-                field.as_ref().dtype
-            ))
-        }));
-    }
-    table.add_row(header);
-    table
 }
